fixed tests

+++ src/main/java/org/example/cafemanager/services/order/impls/OrderServiceImpl.java
@@ -1,4 +1,4 @@
-package org.example.cafemanager.services.order;
+package org.example.cafemanager.services.order.impls;
 
 import org.example.cafemanager.domain.*;
 import org.example.cafemanager.domain.enums.Status;
@@ -9,11 +9,11 @@ import org.example.cafemanager.repositories.OrderRepository;
 import org.example.cafemanager.repositories.ProductsInOrderRepository;
 import org.example.cafemanager.services.exceptions.ChooseAtLeastOneException;
 import org.example.cafemanager.services.exceptions.InstanceNotFoundException;
-import org.example.cafemanager.services.order.contracts.OrderService;
-import org.example.cafemanager.services.product.contracts.ProductService;
-import org.example.cafemanager.services.table.contracts.TableService;
+import org.example.cafemanager.services.order.OrderService;
+import org.example.cafemanager.services.product.ProductService;
+import org.example.cafemanager.services.table.TableService;
 import org.springframework.stereotype.Service;
-
+import org.springframework.util.Assert;
 import javax.transaction.Transactional;
 import javax.validation.constraints.NotNull;
 import java.util.HashSet;
@@ -41,18 +41,19 @@ public class OrderServiceImpl implements OrderService {
 
     @Transactional
     @Override
-    public Order createOrder(OrderDetails orderDetails) {
+    public Order createOrder(@NotNull OrderDetails orderDetails) {
+        Assert.notNull(orderDetails, "OrderDetails can not be null");
         CafeTable table = tableService.getUserAssignedTable(orderDetails.getTableId(), orderDetails.getUser());
-
         if (null == table) {
             throw new InstanceNotFoundException("table");
         }
 
         Set<Product> productsSet = new HashSet<>();
+
         Order order = new Order();
         order.setCafeTable(table);
 
-        Order openedOrder = orderRepository.findOrderByStatus(Status.OPEN);
+        Order openedOrder = orderRepository.findOrderByStatusAndCafeTable(Status.OPEN, order.getCafeTable());
 
         if (null != openedOrder) {
             openedOrder.setStatus(Status.CLOSED);
@@ -62,7 +63,6 @@ public class OrderServiceImpl implements OrderService {
         orderRepository.save(order);
         for (ProductInOrderReq pio : orderDetails.getProdsInOrder()) {
             Product product = productService.findOneById(pio.getProductId());
-
             if (null == product) {
                 continue;
             }
@@ -81,9 +81,17 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     @Transactional
-    public void destroyProductInOrder(Long orderId, Long pioId, User user) {
+    public void deleteProductInOrder(Long orderId, Long productInOrderId, User user) {
         Order order = getOrderByIdAndUser(orderId, user);
-        ProductsInOrder pio = ordersProductInOrder(order, pioId);
+        if (null == order) {
+            throw new InstanceNotFoundException("Order");
+        }
+
+        ProductsInOrder pio = ordersProductInOrder(order, productInOrderId);
+        if (null == pio) {
+            throw new InstanceNotFoundException("Product in order");
+        }
+
         order.removeProductInOrder(pio);
         productsInOrderRepository.delete(pio);
     }
@@ -92,14 +100,27 @@ public class OrderServiceImpl implements OrderService {
     @Transactional
     public void deleteOrder(Long orderId, User user) {
         Order order = getOrderByIdAndUser(orderId, user);
+        if (null == order) {
+            throw new InstanceNotFoundException("Order");
+        }
+
         order.setProductsInOrders(new HashSet<>());
         orderRepository.delete(order);
     }
 
     @Override
-    public ProductsInOrder updateProductInOrder(UpdateProductInOrderDto productUpdate) {
+    public ProductsInOrder updateProductInOrder(@NotNull UpdateProductInOrderDto productUpdate) {
+        Assert.notNull(productUpdate, "UpdateProductInOrderDto can not be null");
         Order order = getOrderByIdAndUser(productUpdate.getOrderId(), productUpdate.getUser());
+        if (null == order) {
+            throw new InstanceNotFoundException("Order");
+        }
+
         ProductsInOrder pio = ordersProductInOrder(order, productUpdate.getPioId());
+        if (null == pio) {
+            throw new InstanceNotFoundException("Product in order");
+        }
+
         pio.setAmount(productUpdate.getAmount());
 
         productsInOrderRepository.save(pio);
@@ -109,8 +130,13 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     @Transactional
-    public Order updateStatus(Long orderId, Status status, User user) {
+    public Order updateOrderStatus(Long orderId, @NotNull Status status, User user) {
+        Assert.notNull(status, "Status can not be null");
         Order order = getOrderByIdAndUser(orderId, user);
+        if (null == order) {
+            throw new InstanceNotFoundException("Order");
+        }
+
         boolean isRequestedStatusIsOpen = status.equals(Status.OPEN);
         if (order.getStatus().equals(status) && isRequestedStatusIsOpen) {
             return order;
@@ -122,7 +148,7 @@ public class OrderServiceImpl implements OrderService {
             return order;
         }
 
-        Order openedOrder = orderRepository.findOrderByStatus(Status.OPEN);
+        Order openedOrder = orderRepository.findOrderByStatusAndCafeTable(Status.OPEN, order.getCafeTable());
 
         if (null != openedOrder) {
             openedOrder.setStatus(Status.CLOSED);
@@ -134,10 +160,11 @@ public class OrderServiceImpl implements OrderService {
         return order;
     }
 
-    private ProductsInOrder ordersProductInOrder(@NotNull Order order, Long pioId) {
+    private ProductsInOrder ordersProductInOrder(@NotNull Order order, Long productInOrderId) {
+        Assert.notNull(order, "Order can not be null");
         ProductsInOrder pio = order.getProductsInOrders()
                 .stream()
-                .filter((productsInOrder) -> productsInOrder.getId().equals(pioId))
+                .filter((productsInOrder) -> productsInOrder.getId().equals(productInOrderId))
                 .findFirst()
                 .orElse(null);
